修复非Windows环境下登录机器人命令出现数字转换问题

+++ src/main/java/me/dreamvoid/miraimc/bukkit/CommandHandler.java
@@ -53,7 +53,7 @@ public class CommandHandler implements CommandExecutor {
                                             sender.sendMessage(ChatColor.translateAlternateColorCodes('&',"&e可用的协议类型: ANDROID_PHONE, ANDROID_PAD, ANDROID_WATCH."));
                                             Protocol = BotConfiguration.MiraiProtocol.ANDROID_PHONE;
                                         }
-                                        mirai.doBotLogin(Integer.parseInt(args[1]),args[2], Protocol);
+                                        mirai.doBotLogin(Long.parseLong(args[1]),args[2], Protocol);
                                     }
                                 }.runTaskAsynchronously(plugin);
                             } else {
